Use hassfest pkg name normalization logic
