+++ script/gen_requirements_all.py
@@ -15,7 +15,7 @@ COMMENT_REQUIREMENTS = (
     "Adafruit_BBIO",
     "avea",  # depends on bluepy
     "avion",
-    "beacontools[scan]",
+    "beacontools",
     "beewi_smartclim",  # depends on bluepy
     "blinkt",
     "bluepy",
@@ -43,6 +43,10 @@ COMMENT_REQUIREMENTS = (
     "tf-models-official",
     "VL53L1X2",
 )
+
+COMMENT_REQUIREMENTS_NORMALIZED = {
+    commented.lower().replace("_", "-") for commented in COMMENT_REQUIREMENTS
+}
 
 IGNORE_PIN = ("colorlog>2.1,<3", "urllib3")
 
@@ -88,6 +92,8 @@ IGNORE_PRE_COMMIT_HOOK_ID = (
     "prettier",
     "python-typing-update",
 )
+
+PACKAGE_REGEX = re.compile(r"^(?:--.+\s)?([-_\.\w\d]+).*==.+$")
 
 
 def has_tests(module: str):
@@ -152,9 +158,24 @@ def gather_recursive_requirements(domain, seen=None):
     return reqs
 
 
+def normalize_package_name(requirement: str) -> str:
+    """Return a normalized package name from a requirement string."""
+    # This function is also used in hassfest.
+    match = PACKAGE_REGEX.search(requirement)
+    if not match:
+        return ""
+
+    # pipdeptree needs lowercase and dash instead of underscore as separator
+    package = match.group(1).lower().replace("_", "-")
+
+    return package
+
+
 def comment_requirement(req):
     """Comment out requirement. Some don't install on all systems."""
-    return any(req.startswith(f"{ign}==") for ign in COMMENT_REQUIREMENTS)
+    return any(
+        normalize_package_name(req) == ign for ign in COMMENT_REQUIREMENTS_NORMALIZED
+    )
 
 
 def gather_modules():
